refactor: keep auto config as module
